Decompose to files and prepare destination classes
